+++ template/loop.php
@@ -11,15 +11,22 @@ $thumb_url = wp_get_attachment_url( get_post_thumbnail_id( $post->ID ) );
         <div class="list-group-item ">
             <div class="row">
                 <div class="col-xs-12 col-sm-6 col-md-6 col-lg-6">
-                    <img width="390" height="440" class="media-object lazy" data-original="<?= $thumb_url; ?>"
+                    <img width="390"
+                         height="440"
+                         class="media-object lazy"
+                         data-original="<?=$thumb_url?>"
                          src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mMMDk76DwAEEgIJ2SbKSQAAAABJRU5ErkJggg=="
-                         alt="<?= sanitize_text_field( get_the_title() ) ?>" itemprop="image">
+                         alt="<?=sanitize_text_field( get_the_title() ) ?>"
+                         itemprop="image"
+                    >
                 </div>
                 <div class="col-xs-12 col-sm-6 col-md-6 col-lg-6">
-                    <h4 class="list-group-item-heading"><a href="<?= get_the_permalink(); ?>"
-                                                           itemprop="name"><?= get_the_title() ?></a></h4>
+                    <h4 class="list-group-item-heading">
+                        <a href="<?=get_the_permalink()?>" itemprop="name"><?=get_the_title()?></a>
+                    </h4>
                     <p class="list-group-item-text" itemprop="description">
-						<?= get_short_description( get_the_content(), 50 ); ?>
+						<?=get_short_description( get_the_content(), 50 )?>
+                    </p>
                     <table class="table">
                         <tbody>
                         <tr>
@@ -32,12 +39,6 @@ $thumb_url = wp_get_attachment_url( get_post_thumbnail_id( $post->ID ) );
                             <td>Год выхода:</td>
                             <td itemprop="copyrightYear">
 								<?= get_post_meta( $post->ID, 'year', true ); ?>
-                            </td>
-                        </tr>
-                        <tr>
-                            <td>Жанр:</td>
-                            <td itemprop="genre">
-								<?php the_category() ?>
                             </td>
                         </tr>
                         <tr>
@@ -67,19 +68,16 @@ $thumb_url = wp_get_attachment_url( get_post_thumbnail_id( $post->ID ) );
 								<?= get_post_meta( $post->ID, 'format', true ); ?>
                             </td>
                         </tr>
-                        <tr>
-                            <td>Cтраниц:</td>
-                            <td itemprop="numberOfPages">
-								<?= get_post_meta( $post->ID, 'number_page', true ); ?>
-                            </td>
-                        </tr>
                         </tbody>
                     </table>
-                    </p>
                 </div>
                 <div class="next-reed">
-                    <p><a href="<?= get_the_permalink(); ?>"
-                          class="btn navbar-btn btn-info navbar-right">Дальше</a></p>
+                    <p>
+                        <a
+                            href="<?=get_the_permalink()?>"
+                            class="btn navbar-btn btn-info navbar-right">Дальше
+                        </a>
+                    </p>
                 </div>
             </div>
         </div>